Minor changes and fixes to my code, will need to do this again later tonight
