+++ AdminSystem/StockDataEntry.aspx.cs
@@ -10,7 +10,12 @@ public partial class _1_DataEntry : System.Web.UI.Page
 {
     protected void Page_Load(object sender, EventArgs e)
     {
-
+        //create new instance of class 
+        clsStock AnStock = new clsStock();
+        //get data from the session object
+        AnStock = (clsStock)Session["AnStock"];
+        //display the game number on the page
+        Response.Write(AnStock.GameNumber + "<br />" + AnStock.GameNumber + "<br />" + "£" + AnStock.DateAdded + "<br />" + AnStock.AgeRating + "<br />" + AnStock.AgeRating);
     }
 
     protected void btnOK_Click(object sender, EventArgs e)
